% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{addStatement}
\alias{addStatement}
\title{Add a Statement object to the Model}
\usage{
addStatement(.Object, statement)
}
\arguments{
\item{.Object}{a Model object}

\item{statement}{the Statement that will be added}
}
\description{
Add a Statement object to the Model
}

