% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Query.R
\name{executeQuery}
\alias{executeQuery}
\title{Execute a query}
\usage{
executeQuery(.Object, model)
}
\arguments{
\item{.Object}{a Query object}

\item{model}{a Model object}
}
\value{
a QueryResults object
}
\description{
The initialize query is executed and the result is returned as a QueryResult object
}

