% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Serializer.R
\docType{methods}
\name{freeSerializer,Serializer-method}
\alias{freeSerializer,Serializer-method}
\title{Free memory used by a librdf serializer}
\usage{
\S4method{freeSerializer}{Serializer}(.Object)
}
\arguments{
\item{.Object}{a Serializer object}
}
\description{
Free memory used by a librdf serializer
}

