% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Serializer.R
\name{freeSerializer}
\alias{freeSerializer}
\title{Free memory used by a librdf serializer}
\usage{
freeSerializer(.Object)
}
\arguments{
\item{.Object}{a Serializer object}
}
\description{
Free memory used by a librdf serializer
}

