% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statement.R
\docType{methods}
\name{freeStatement}
\alias{freeStatement}
\alias{freeStatement,Statement-method}
\title{Free memory used by a librdf statement}
\usage{
freeStatement(.Object)

\S4method{freeStatement}{Statement}(.Object)
}
\arguments{
\item{.Object}{a Statement object}
}
\description{
Free memory used by a librdf statement
}
\details{
After this method is called, the Statement object is no longer usable and should
be deleted  \code{"rm(statement)"} and a new object created.
}

