% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Node.R
\name{getBlankNodeId}
\alias{getBlankNodeId}
\title{Get the blank identifier that has been assigned for a specified Node object}
\usage{
getBlankNodeId(.Object)
}
\arguments{
\item{.Object}{a Node object}
}
\value{
a blank node identifier
}
\description{
Get the blank identifier that has been assigned for a specified Node object
}
\details{
When a Node object is initialized with no value specified, i.e. node <- Node(""),
a blank node is created and a locally unique identifier is generated by librdf. This method
retrieves this identifier and returns in to the caller.
}

