% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/QueryResults.R
\name{getNextResult}
\alias{getNextResult}
\title{Get the next query result}
\usage{
getNextResult(.Object)
}
\arguments{
\item{.Object}{a QueryResults object}
}
\description{
The next query result is returned
}

