% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Node.R
\name{getNodeType}
\alias{getNodeType}
\title{Determine the node type and return as a string}
\usage{
getNodeType(.Object)
}
\arguments{
\item{.Object}{a Node object}
}
\value{
a character vector containing the Node type
}
\description{
A Node has a type that is assigned at initialization and can have one of the following values:
'resource', 'literal', 'blank' and 'unknown'.
}

