% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Query.R
\name{getQueryResultLimit}
\alias{getQueryResultLimit}
\title{Get the query result limit}
\usage{
getQueryResultLimit(.Object)
}
\arguments{
\item{.Object}{a Query object}
}
\value{
the query result limit. If a limit is set then the value will be >= 0. If the value is < 0, no limit is set
}
\description{
Get the query result limit
}

