% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\docType{methods}
\name{getQueryResultLimit}
\alias{getQueryResultLimit}
\alias{getQueryResultLimit,Query-method}
\title{Get the query result limit}
\usage{
getQueryResultLimit(.Object)

\S4method{getQueryResultLimit}{Query}(.Object)
}
\arguments{
\item{.Object}{a Query object}
}
\value{
the query result limit. If a limit is set then the value will be >= 0. If the value is < 0, no limit is set
}
\description{
Get the query result limit
}
