% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initialize,Model-method}
\alias{initialize,Model-method}
\title{Constructor for a Model object.}
\usage{
\S4method{initialize}{Model}(.Object, world, storage, options)
}
\value{
the World object
}
\description{
Constructor for a Model object.
}

