% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initialize,Node-method}
\alias{initialize,Node-method}
\title{Initialize a Node object.}
\usage{
\S4method{initialize}{Node}(.Object, world, literal = as.character(NA))
}
\arguments{
\item{.Object}{the Node object to be initialized}

\item{world}{a World object}

\item{literal}{a literal character value to be used as the vale of the node}
}
\value{
the Node object
}
\description{
Initialize a Node object.
}

