% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initialize,Statement-method}
\alias{initialize,Statement-method}
\title{Construct a Statement object.}
\usage{
\S4method{initialize}{Statement}(.Object, world, subject, predicate, object)
}
\arguments{
\item{world}{a World object}

\item{subject}{a Node object}

\item{predicate}{a Node object}

\item{object}{a Node object}
}
\value{
the Statement object
}
\description{
Construct a Statement object.
}

