% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initialize,Storage-method}
\alias{initialize,Storage-method}
\title{Constructor for Storage object}
\usage{
\S4method{initialize}{Storage}(.Object, world, type = "hashes", name = "",
  options = "hash-type='memory'")
}
\arguments{
\item{world}{the World object}

\item{type}{the Redland storage type}

\item{name}{storage instance name}

\item{options}{storage options}
}
\value{
the Storage object
}
\description{
Constructor for Storage object
}

