% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is.null.externalptr}
\alias{is.null.externalptr}
\title{Determine whether an externalptr object is NULL.}
\usage{
is.null.externalptr(pointer)
}
\arguments{
\item{pointer}{externalptr to be checked for NULL value}
}
\value{
logical TRUE if the pointer is NULL, otherwise FALSE
}
\description{
The pointer is treated as an externalptr and checked via a call in C to see if it is NULL.
}
