% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/redland.R
\docType{methods}
\name{length,SWIGArray-method}
\alias{length,SWIGArray-method}
\title{Return length of a SWIGArray}
\usage{
\S4method{length}{SWIGArray}(x)
}
\arguments{
\item{x}{the SWIGArray}
}
\description{
Return length of a SWIGArray
}

