% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_copyright_string}
\alias{librdf_copyright_string}
\title{Copyright string (multiple lines).}
\usage{
librdf_copyright_string ( .copy )
}
\arguments{

\item{.copy}{NA}

}
\value{
character
}
\description{
Copyright string (multiple lines).
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
