\name{librdf_copyright_string_get}
\alias{librdf_copyright_string_get}
\title{Return Redland RDF copyright string}
\usage{
librdf_copyright_string_get (.copy)
}
\arguments{

\item{.copy}{logical}

}
\value{
character
}
\description{
Return the Redland RDF copyright
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
