% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_digest_final}
\alias{librdf_digest_final}
\title{Finish the digesting of data.}
\usage{
librdf_digest_final ( digest )
}
\arguments{

\item{digest}{the digest ("_p_librdf_digest_s")}

}
\value{
void
}
\description{
Finish the digesting of data.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
