% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_digest_to_string}
\alias{librdf_digest_to_string}
\title{Get a string representation of the digest object.}
\usage{
librdf_digest_to_string ( digest )
}
\arguments{

\item{digest}{the digest ("_p_librdf_digest_s")}

}
\value{
character
}
\description{
Get a string representation of the digest object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
