% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_digest_update}
\alias{librdf_digest_update}
\title{Add more data to the librdf_digest object.}
\usage{
librdf_digest_update ( digest,
	buf,
	length )
}
\arguments{

\item{digest}{the digest ("_p_librdf_digest_s")}

\item{buf}{the data buffer ("character")}

\item{length}{the length of the data ("integer")}

}
\value{
void
}
\description{
Add more data to the librdf_digest object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
