% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_digest_update_string}
\alias{librdf_digest_update_string}
\title{Add a string to the librdf_digest object.}
\usage{
librdf_digest_update_string ( digest,
	string )
}
\arguments{

\item{digest}{the digest ("_p_librdf_digest_s")}

\item{string}{string to add ("character")}

}
\value{
void
}
\description{
Add a string to the librdf_digest object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
