% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_free_digest}
\alias{librdf_free_digest}
\title{Destructor - destroy a librdf_digest object.}
\usage{
librdf_free_digest ( digest )
}
\arguments{

\item{digest}{the digest ("_p_librdf_digest_s")}

}
\value{
void
}
\description{
Destructor - destroy a librdf_digest object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
