% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_free_hash}
\alias{librdf_free_hash}
\title{Destructor - destroy a librdf_hash object.}
\usage{
librdf_free_hash ( hash )
}
\arguments{

\item{hash}{hash object ("_p_librdf_hash_s")}

}
\value{
void
}
\description{
Destructor - destroy a librdf_hash object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
