% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_free_iterator}
\alias{librdf_free_iterator}
\title{Destructor - destroy a librdf_iterator object.}
\usage{
librdf_free_iterator ( s_arg1 )
}
\arguments{

\item{s_arg1}{the librdf_iterator object ("_p_librdf_iterator_s")}

}
\value{
void
}
\description{
Destructor - destroy a librdf_iterator object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
