% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_free_model}
\alias{librdf_free_model}
\title{Destructor - Destroy a librdf_model object.}
\usage{
librdf_free_model ( model )
}
\arguments{

\item{model}{librdf_model model to destroy ("_p_librdf_model_s")}

}
\value{
void
}
\description{
Destructor - Destroy a librdf_model object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
