% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_free_node}
\alias{librdf_free_node}
\title{Destructor - destroy an librdf_node object.}
\usage{
librdf_free_node ( r )
}
\arguments{

\item{r}{librdf_node object ("_p_librdf_node_s")}

}
\value{
void
}
\description{
Destructor - destroy an librdf_node object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
