% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_free_parser}
\alias{librdf_free_parser}
\title{Destructor - destroys a librdf_parser object.}
\usage{
librdf_free_parser ( parser )
}
\arguments{

\item{parser}{the parser ("_p_librdf_parser_s")}

}
\value{
void
}
\description{
Destructor - destroys a librdf_parser object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
