% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_free_query}
\alias{librdf_free_query}
\title{Destructor - destroy a librdf_query object.}
\usage{
librdf_free_query ( query )
}
\arguments{

\item{query}{librdf_query object ("_p_librdf_query")}

}
\value{
void
}
\description{
Destructor - destroy a librdf_query object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
