% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_free_query_results}
\alias{librdf_free_query_results}
\title{Destructor - destroy a librdf_query_results object.}
\usage{
librdf_free_query_results ( query_results )
}
\arguments{

\item{query_results}{librdf_query_results object ("_p_librdf_query_results")}

}
\value{
void
}
\description{
Destructor - destroy a librdf_query_results object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
