% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_free_serializer}
\alias{librdf_free_serializer}
\title{Destructor - destroys a librdf_serializer object.}
\usage{
librdf_free_serializer ( serializer )
}
\arguments{

\item{serializer}{the serializer ("_p_librdf_serializer_s")}

}
\value{
void
}
\description{
Destructor - destroys a librdf_serializer object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
