% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_free_statement}
\alias{librdf_free_statement}
\title{Destructor - destroy a librdf_statement.}
\usage{
librdf_free_statement ( statement )
}
\arguments{

\item{statement}{librdf_statement object ("_p_librdf_statement_s")}

}
\value{
void
}
\description{
Destructor - destroy a librdf_statement.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
