% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_free_storage}
\alias{librdf_free_storage}
\title{Destructor - destroy a librdf_storage object.}
\usage{
librdf_free_storage ( storage )
}
\arguments{

\item{storage}{librdf_storage object ("_p_librdf_storage_s")}

}
\value{
void
}
\description{
Destructor - destroy a librdf_storage object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
