% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_free_stream}
\alias{librdf_free_stream}
\title{Destructor - destroy an libdf_stream object.}
\usage{
librdf_free_stream ( stream )
}
\arguments{

\item{stream}{librdf_stream object ("_p_librdf_stream_s")}

}
\value{
void
}
\description{
Destructor - destroy an libdf_stream object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
