% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_free_uri}
\alias{librdf_free_uri}
\title{Destructor - destroy a librdf_uri object.}
\usage{
librdf_free_uri ( uri )
}
\arguments{

\item{uri}{librdf_uri object ("_p_librdf_uri_s")}

}
\value{
void
}
\description{
Destructor - destroy a librdf_uri object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
