% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_free_world}
\alias{librdf_free_world}
\title{Terminate the library and frees all allocated resources.}
\usage{
librdf_free_world ( world )
}
\arguments{

\item{world}{redland world object ("_p_librdf_world_s")}

}
\value{
void
}
\description{
Terminate the library and frees all allocated resources.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
