% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_hash_to_string}
\alias{librdf_hash_to_string}
\title{Format the hash as a string, suitable for parsing by librdf_hash_from_string.}
\usage{
librdf_hash_to_string ( hash,
	filter )
}
\arguments{

\item{hash}{librdf_hash object ("_p_librdf_hash_s")}

\item{filter}{NULL terminated list of keys to ignore ("_p_p_char")}

}
\value{
character
}
\description{
Format the hash as a string, suitable for parsing by librdf_hash_from_string.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
