\name{librdf_internal_test_warning}
\alias{librdf_internal_test_warning}
\title{For internal testing, not part of public API}
\usage{
librdf_internal_test_warning ( world )
}
\arguments{

\item{world}{librdf_world ("_p_librdf_world_s")}

}
\value{
void
}
\description{
This funciton is for internal testing of the Redland software and is not part of the public API.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
