% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_iterator_end}
\alias{librdf_iterator_end}
\title{Test if the iterator has finished.}
\usage{
librdf_iterator_end ( iterator,
	.copy )
}
\arguments{

\item{iterator}{the librdf_iterator object ("_p_librdf_iterator_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Test if the iterator has finished.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
