% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_iterator_get_context}
\alias{librdf_iterator_get_context}
\title{Get the context of the current object on the iterator.}
\usage{
librdf_iterator_get_context ( iterator )
}
\arguments{

\item{iterator}{the librdf_iterator object ("_p_librdf_iterator_s")}

}
\value{
_p_librdf_node_s
}
\description{
Get the context of the current object on the iterator.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
