% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_iterator_next}
\alias{librdf_iterator_next}
\title{Move to the next iterator element.}
\usage{
librdf_iterator_next ( iterator,
	.copy )
}
\arguments{

\item{iterator}{the librdf_iterator object ("_p_librdf_iterator_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Move to the next iterator element.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
