% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_log_message_code}
\alias{librdf_log_message_code}
\title{Retrieve error code from log message.}
\usage{
librdf_log_message_code ( message,
	.copy )
}
\arguments{

\item{message}{log message ("_p_librdf_log_message")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Retrieve error code from log message.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
