% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_log_message_locator}
\alias{librdf_log_message_locator}
\title{Retrieve locator of log entry.}
\usage{
librdf_log_message_locator ( message )
}
\arguments{

\item{message}{log message ("_p_librdf_log_message")}

}
\value{
_p_raptor_locator
}
\description{
Retrieve locator of log entry.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
