% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_add}
\alias{librdf_model_add}
\title{Create and add a new statement about a resource to the model.}
\usage{
librdf_model_add ( model,
	subject,
	predicate,
	object,
	.copy )
}
\arguments{

\item{model}{model object ("_p_librdf_model_s")}

\item{subject}{librdf_node of subject ("_p_librdf_node_s")}

\item{predicate}{librdf_node of predicate ("_p_librdf_node_s")}

\item{object}{librdf_node of object (literal or resource) ("_p_librdf_node_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Create and add a new statement about a resource to the model.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
