% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_add_statement}
\alias{librdf_model_add_statement}
\title{Add a statement to the model.}
\usage{
librdf_model_add_statement ( model,
	statement,
	.copy )
}
\arguments{

\item{model}{model object ("_p_librdf_model_s")}

\item{statement}{statement object ("_p_librdf_statement_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Add a statement to the model.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
