% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_as_stream}
\alias{librdf_model_as_stream}
\title{List the model contents as a stream of statements.}
\usage{
librdf_model_as_stream ( model )
}
\arguments{

\item{model}{the model object ("_p_librdf_model_s")}

}
\value{
_p_librdf_stream_s
}
\description{
List the model contents as a stream of statements.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
