% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_contains_context}
\alias{librdf_model_contains_context}
\title{Check for a context in the model.}
\usage{
librdf_model_contains_context ( model,
	context,
	.copy )
}
\arguments{

\item{model}{the model object ("_p_librdf_model_s")}

\item{context}{the contest ("_p_librdf_node_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Check for a context in the model.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
