% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_context_as_stream}
\alias{librdf_model_context_as_stream}
\title{List all statements in a model context.}
\usage{
librdf_model_context_as_stream ( model,
	context )
}
\arguments{

\item{model}{librdf_model object ("_p_librdf_model_s")}

\item{context}{librdf_node context ("_p_librdf_node_s")}

}
\value{
_p_librdf_stream_s
}
\description{
List all statements in a model context.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
