% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_context_remove_statement}
\alias{librdf_model_context_remove_statement}
\title{Remove a statement from a model in a context.}
\usage{
librdf_model_context_remove_statement ( model,
	context,
	statement,
	.copy )
}
\arguments{

\item{model}{librdf_model object ("_p_librdf_model_s")}

\item{context}{librdf_node context ("_p_librdf_node_s")}

\item{statement}{librdf_statement statement ("_p_librdf_statement_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Remove a statement from a model in a context.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
