% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_context_remove_statements}
\alias{librdf_model_context_remove_statements}
\title{Remove statements from a model with the given context.}
\usage{
librdf_model_context_remove_statements ( model,
	context,
	.copy )
}
\arguments{

\item{model}{librdf_model object ("_p_librdf_model_s")}

\item{context}{librdf_node context ("_p_librdf_node_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Remove statements from a model with the given context.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
