% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_find_statements}
\alias{librdf_model_find_statements}
\title{Find matching statements in the model.}
\usage{
librdf_model_find_statements ( model,
	statement )
}
\arguments{

\item{model}{the model object ("_p_librdf_model_s")}

\item{statement}{the partial statement to match ("_p_librdf_statement_s")}

}
\value{
_p_librdf_stream_s
}
\description{
Find matching statements in the model.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
