% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_find_statements_in_context}
\alias{librdf_model_find_statements_in_context}
\title{Search the model for matching statements in a given context.}
\usage{
librdf_model_find_statements_in_context ( model,
	statement,
	inNodeOrNull )
}
\arguments{

\item{model}{librdf_model object ("_p_librdf_model_s")}

\item{statement}{librdf_statement partial statement to find ("_p_librdf_statement_s")}

\item{inNodeOrNull}{context librdf_node (or NULL) ("_p_librdf_node_s")}

}
\value{
_p_librdf_stream_s
}
\description{
Search the model for matching statements in a given context.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
