% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_get_arcs}
\alias{librdf_model_get_arcs}
\title{Return the arcs (predicates) of an arc in an RDF graph given source (subject) and target (object).}
\usage{
librdf_model_get_arcs ( model,
	source,
	target )
}
\arguments{

\item{model}{librdf_model object ("_p_librdf_model_s")}

\item{source}{librdf_node source ("_p_librdf_node_s")}

\item{target}{librdf_node target ("_p_librdf_node_s")}

}
\value{
_p_librdf_iterator_s
}
\description{
Return the arcs (predicates) of an arc in an RDF graph given source (subject) and target (object).
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
