% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_get_arcs_out}
\alias{librdf_model_get_arcs_out}
\title{Return the properties pointing from the given resource.}
\usage{
librdf_model_get_arcs_out ( model,
	node )
}
\arguments{

\item{model}{librdf_model object ("_p_librdf_model_s")}

\item{node}{librdf_node resource node ("_p_librdf_node_s")}

}
\value{
_p_librdf_iterator_s
}
\description{
Return the properties pointing from the given resource.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
