% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_get_contexts}
\alias{librdf_model_get_contexts}
\title{Return the list of contexts in the graph.}
\usage{
librdf_model_get_contexts ( model )
}
\arguments{

\item{model}{librdf_model object ("_p_librdf_model_s")}

}
\value{
_p_librdf_iterator_s
}
\description{
Return the list of contexts in the graph.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
