% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_get_feature}
\alias{librdf_model_get_feature}
\title{Get the value of a graph feature .}
\usage{
librdf_model_get_feature ( model,
	feature )
}
\arguments{

\item{model}{librdf_model object ("_p_librdf_model_s")}

\item{feature}{librdf_uri feature property ("_p_librdf_uri_s")}

}
\value{
_p_librdf_node_s
}
\description{
Get the value of a graph feature .
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
