% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_get_source}
\alias{librdf_model_get_source}
\title{Return one source (subject) of arc in an RDF graph given arc (predicate) and target (object).}
\usage{
librdf_model_get_source ( model,
	arc,
	target )
}
\arguments{

\item{model}{librdf_model object ("_p_librdf_model_s")}

\item{arc}{librdf_node arc ("_p_librdf_node_s")}

\item{target}{librdf_node target ("_p_librdf_node_s")}

}
\value{
_p_librdf_node_s
}
\description{
Return one source (subject) of arc in an RDF graph given arc (predicate) and target (object).
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
