% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_get_targets}
\alias{librdf_model_get_targets}
\title{Return the targets (objects) of an arc in an RDF graph given source (subject) and arc (predicate).}
\usage{
librdf_model_get_targets ( model,
	source,
	arc )
}
\arguments{

\item{model}{librdf_model object ("_p_librdf_model_s")}

\item{source}{librdf_node source ("_p_librdf_node_s")}

\item{arc}{librdf_node arc ("_p_librdf_node_s")}

}
\value{
_p_librdf_iterator_s
}
\description{
Return the targets (objects) of an arc in an RDF graph given source (subject) and arc (predicate).
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
