% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_has_arc_in}
\alias{librdf_model_has_arc_in}
\title{Check if a node has a given property pointing to it.}
\usage{
librdf_model_has_arc_in ( model,
	node,
	property,
	.copy )
}
\arguments{

\item{model}{librdf_model object ("_p_librdf_model_s")}

\item{node}{librdf_node resource node ("_p_librdf_node_s")}

\item{property}{librdf_node property node ("_p_librdf_node_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Check if a node has a given property pointing to it.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
