% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_load}
\alias{librdf_model_load}
\title{Load content from a URI into the model.}
\usage{
librdf_model_load ( model,
	uri,
	name,
	mime_type,
	type_uri,
	.copy )
}
\arguments{

\item{model}{librdf_model object ("_p_librdf_model_s")}

\item{uri}{the URI to read the content ("_p_librdf_uri_s")}

\item{name}{the name of the parser (or NULL) ("character")}

\item{mime_type}{the MIME type of the syntax (NULL if not used) ("character")}

\item{type_uri}{URI identifying the syntax (NULL if not used) ("_p_librdf_uri_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Load content from a URI into the model.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
