% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_query_execute}
\alias{librdf_model_query_execute}
\title{Execute a query against the model.}
\usage{
librdf_model_query_execute ( model,
	query )
}
\arguments{

\item{model}{librdf_model object ("_p_librdf_model_s")}

\item{query}{librdf_query object ("_p_librdf_query")}

}
\value{
_p_librdf_query_results
}
\description{
Execute a query against the model.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
