% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_remove_statement}
\alias{librdf_model_remove_statement}
\title{Remove a known statement from the model.}
\usage{
librdf_model_remove_statement ( model,
	statement,
	.copy )
}
\arguments{

\item{model}{the model object ("_p_librdf_model_s")}

\item{statement}{the statement ("_p_librdf_statement_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Remove a known statement from the model.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
