% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_size}
\alias{librdf_model_size}
\title{Get the number of statements in the model.}
\usage{
librdf_model_size ( model,
	.copy )
}
\arguments{

\item{model}{librdf_model object ("_p_librdf_model_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Get the number of statements in the model.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
