% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_sync}
\alias{librdf_model_sync}
\title{Synchronise the model to the model implementation.}
\usage{
librdf_model_sync ( model )
}
\arguments{

\item{model}{librdf_model object ("_p_librdf_model_s")}

}
\value{
void
}
\description{
Synchronise the model to the model implementation.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
