% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_to_string}
\alias{librdf_model_to_string}
\title{Write serialized model to a string.}
\usage{
librdf_model_to_string ( model,
	uri,
	name,
	mime_type,
	inUriOrNull )
}
\arguments{

\item{model}{librdf_model object ("_p_librdf_model_s")}

\item{uri}{base URI to use in serializing (or NULL if not used) ("_p_librdf_uri_s")}

\item{name}{the name of the serializer (or NULL for default) ("character")}

\item{mime_type}{the MIME type of the syntax (NULL if not used) ("character")}

\item{inUriOrNull}{URI identifying the syntax (NULL if not used) ("_p_librdf_uri_s")}

}
\value{
character
}
\description{
Write serialized model to a string.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
