% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_model_transaction_start}
\alias{librdf_model_transaction_start}
\title{Start a transaction}
\usage{
librdf_model_transaction_start ( model,
	.copy )
}
\arguments{

\item{model}{the model object ("_p_librdf_model_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Start a transaction
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
