% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_digest}
\alias{librdf_new_digest}
\title{Constructor - create a new librdf_digest object.}
\usage{
librdf_new_digest ( world,
	name )
}
\arguments{

\item{world}{redland world object ("_p_librdf_world_s")}

\item{name}{the digest name to use to create this digest ("character")}

}
\value{
_p_librdf_digest_s
}
\description{
Constructor - create a new librdf_digest object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
