% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_hash_from_string}
\alias{librdf_new_hash_from_string}
\title{Constructor - create a new librdf_hash object from a string.}
\usage{
librdf_new_hash_from_string ( world,
	name,
	string )
}
\arguments{

\item{world}{redland world object ("_p_librdf_world_s")}

\item{name}{hash name ("character")}

\item{string}{hash encoded as a string ("character")}

}
\value{
_p_librdf_hash_s
}
\description{
Constructor - create a new librdf_hash object from a string.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
