% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_model}
\alias{librdf_new_model}
\title{Constructor - create a new storage librdf_model object.}
\usage{
librdf_new_model ( world,
	storage,
	options_string )
}
\arguments{

\item{world}{redland world object ("_p_librdf_world_s")}

\item{storage}{librdf_storage to use ("_p_librdf_storage_s")}

\item{options_string}{options to initialise model ("character")}

}
\value{
_p_librdf_model_s
}
\description{
Constructor - create a new storage librdf_model object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
