% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_model_from_model}
\alias{librdf_new_model_from_model}
\title{Copy constructor - create a new librdf_model from an existing one.}
\usage{
librdf_new_model_from_model ( model )
}
\arguments{

\item{model}{the existing librdf_model ("_p_librdf_model_s")}

}
\value{
_p_librdf_model_s
}
\description{
Copy constructor - create a new librdf_model from an existing one.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
