% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_model_with_options}
\alias{librdf_new_model_with_options}
\title{Constructor - Create a new librdf_model with storage.}
\usage{
librdf_new_model_with_options ( world,
	storage,
	options )
}
\arguments{

\item{world}{redland world object ("_p_librdf_world_s")}

\item{storage}{librdf_storage storage to use ("_p_librdf_storage_s")}

\item{options}{librdf_hash of options to use ("_p_librdf_hash_s")}

}
\value{
_p_librdf_model_s
}
\description{
Constructor - Create a new librdf_model with storage.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
