% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_node_from_blank_identifier}
\alias{librdf_new_node_from_blank_identifier}
\title{Constructor - create a new blank node librdf_node object from a blank node identifier.}
\usage{
librdf_new_node_from_blank_identifier ( world,
	inStrOrNull )
}
\arguments{

\item{world}{redland world object ("_p_librdf_world_s")}

\item{inStrOrNull}{UTF-8 encoded blank node identifier or NULL ("character")}

}
\value{
_p_librdf_node_s
}
\description{
Constructor - create a new blank node librdf_node object from a blank node identifier.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
