% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_node_from_literal}
\alias{librdf_new_node_from_literal}
\title{Constructor - create a new literal librdf_node object.}
\usage{
librdf_new_node_from_literal ( world,
	string,
	inStrOrNull,
	is_wf_xml )
}
\arguments{

\item{world}{redland world object ("_p_librdf_world_s")}

\item{string}{literal UTF-8 encoded string value ("character")}

\item{inStrOrNull}{literal XML language (or NULL, empty string) ("character")}

\item{is_wf_xml}{non 0 if literal is XML ("integer")}

}
\value{
_p_librdf_node_s
}
\description{
Constructor - create a new literal librdf_node object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
