% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_node_from_node}
\alias{librdf_new_node_from_node}
\title{Copy constructor - create a new librdf_node object from an existing librdf_node object.}
\usage{
librdf_new_node_from_node ( node )
}
\arguments{

\item{node}{librdf_node object to copy ("_p_librdf_node_s")}

}
\value{
_p_librdf_node_s
}
\description{
Copy constructor - create a new librdf_node object from an existing librdf_node object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
