% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_node_from_normalised_uri_string}
\alias{librdf_new_node_from_normalised_uri_string}
\title{Constructor - create a new librdf_node object from a UTF-8 encoded URI string normalised to a new base URI.}
\usage{
librdf_new_node_from_normalised_uri_string ( world,
	uri_string,
	source_uri,
	base_uri )
}
\arguments{

\item{world}{redland world object ("_p_librdf_world_s")}

\item{uri_string}{UTF-8 encoded string representing a URI ("character")}

\item{source_uri}{source URI ("_p_librdf_uri_s")}

\item{base_uri}{base URI ("_p_librdf_uri_s")}

}
\value{
_p_librdf_node_s
}
\description{
Constructor - create a new librdf_node object from a UTF-8 encoded URI string normalised to a new base URI.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
