% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_node_from_uri}
\alias{librdf_new_node_from_uri}
\title{Constructor - create a new resource librdf_node object with a given URI.}
\usage{
librdf_new_node_from_uri ( world,
	uri )
}
\arguments{

\item{world}{redland world object ("_p_librdf_world_s")}

\item{uri}{librdf_uri object ("_p_librdf_uri_s")}

}
\value{
_p_librdf_node_s
}
\description{
Constructor - create a new resource librdf_node object with a given URI.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
