% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_node_from_uri_string}
\alias{librdf_new_node_from_uri_string}
\title{Constructor - create a new librdf_node object from a URI string.}
\usage{
librdf_new_node_from_uri_string ( world,
	string )
}
\arguments{

\item{world}{redland world object ("_p_librdf_world_s")}

\item{string}{string representing a URI ("character")}

}
\value{
_p_librdf_node_s
}
\description{
Constructor - create a new librdf_node object from a URI string.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
