% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_query}
\alias{librdf_new_query}
\title{Constructor - create a new librdf_query object.}
\usage{
librdf_new_query ( world,
	name,
	uri,
	query_string,
	base_uri )
}
\arguments{

\item{world}{redland world object ("_p_librdf_world_s")}

\item{name}{the name identifying the query language ("character")}

\item{uri}{the URI identifying the query language (or NULL) ("_p_librdf_uri_s")}

\item{query_string}{the query string ("character")}

\item{base_uri}{the base URI of the query string (or NULL) ("_p_librdf_uri_s")}

}
\value{
_p_librdf_query
}
\description{
Constructor - create a new librdf_query object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
