% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_query_from_query}
\alias{librdf_new_query_from_query}
\title{Copy constructor - create a new librdf_query object from an existing one}
\usage{
librdf_new_query_from_query ( old_query )
}
\arguments{

\item{old_query}{the existing query librdf_query to use ("_p_librdf_query")}

}
\value{
_p_librdf_query
}
\description{
Copy constructor - create a new librdf_query object from an existing one
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
