% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_serializer}
\alias{librdf_new_serializer}
\title{Constructor - create a new librdf_serializer object.}
\usage{
librdf_new_serializer ( world,
	name,
	mime_type,
	type_uri )
}
\arguments{

\item{world}{redland world object ("_p_librdf_world_s")}

\item{name}{the serializer factory name (or NULL or empty string if don't care) ("character")}

\item{mime_type}{the MIME type of the syntax (NULL if not used) ("character")}

\item{type_uri}{URI of syntax (NULL if not used) ("_p_librdf_uri_s")}

}
\value{
_p_librdf_serializer_s
}
\description{
Constructor - create a new librdf_serializer object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
