% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_statement}
\alias{librdf_new_statement}
\title{Constructor - create a new empty librdf_statement.}
\usage{
librdf_new_statement ( world )
}
\arguments{

\item{world}{redland world object ("_p_librdf_world_s")}

}
\value{
_p_librdf_statement_s
}
\description{
Constructor - create a new empty librdf_statement.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
