% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_statement_from_statement}
\alias{librdf_new_statement_from_statement}
\title{Copy constructor - create a new librdf_statement from an existing librdf_statement.
Creates a deep copy - changes to original statement nodes are not reflected in the copy.}
\usage{
librdf_new_statement_from_statement ( statement )
}
\arguments{

\item{statement}{librdf_statement to copy ("_p_librdf_statement_s")}

}
\value{
_p_librdf_statement_s
}
\description{
Copy constructor - create a new librdf_statement from an existing librdf_statement.
Creates a deep copy - changes to original statement nodes are not reflected in the copy.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
