% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_storage}
\alias{librdf_new_storage}
\title{Constructor - create a new librdf_storage object.}
\usage{
librdf_new_storage ( world,
	storage_name,
	name,
	options_string )
}
\arguments{

\item{world}{redland world object ("_p_librdf_world_s")}

\item{storage_name}{the storage factory name ("character")}

\item{name}{an identifier for the storage ("character")}

\item{options_string}{options to initialise storage ("character")}

}
\value{
_p_librdf_storage_s
}
\description{
Constructor - create a new librdf_storage object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
