% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_storage_from_storage}
\alias{librdf_new_storage_from_storage}
\title{Copy constructor - create a new librdf_storage object from an existing one}
\usage{
librdf_new_storage_from_storage ( old_storage )
}
\arguments{

\item{old_storage}{the existing storage librdf_storage to use ("_p_librdf_storage_s")}

}
\value{
_p_librdf_storage_s
}
\description{
Copy constructor - create a new librdf_storage object from an existing one
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
