% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_uri}
\alias{librdf_new_uri}
\title{Constructor - create a new librdf_uri object from a URI string.}
\usage{
librdf_new_uri ( world,
	string )
}
\arguments{

\item{world}{redland world object ("_p_librdf_world_s")}

\item{string}{URI in string form ("character")}

}
\value{
_p_librdf_uri_s
}
\description{
Constructor - create a new librdf_uri object from a URI string.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
