% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_uri_from_filename}
\alias{librdf_new_uri_from_filename}
\title{Constructor - create a new librdf_uri object from a filename.}
\usage{
librdf_new_uri_from_filename ( world,
	filename )
}
\arguments{

\item{world}{Redland librdf_world object ("_p_librdf_world_s")}

\item{filename}{filename ("character")}

}
\value{
_p_librdf_uri_s
}
\description{
Constructor - create a new librdf_uri object from a filename.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
