% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_uri_from_uri}
\alias{librdf_new_uri_from_uri}
\title{Copy constructor - create a new librdf_uri object from an existing librdf_uri object.}
\usage{
librdf_new_uri_from_uri ( uri )
}
\arguments{

\item{uri}{librdf_uri object ("_p_librdf_uri_s")}

}
\value{
_p_librdf_uri_s
}
\description{
Copy constructor - create a new librdf_uri object from an existing librdf_uri object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
