% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_new_world}
\alias{librdf_new_world}
\title{Create a new Redland execution environment.}
\usage{
librdf_new_world (  )
}
\value{
_p_librdf_world_s
}
\description{
Create a new Redland execution environment.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
