% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_node_equals}
\alias{librdf_node_equals}
\title{Compare two librdf_node objects for equality.}
\usage{
librdf_node_equals ( first_node,
	second_node,
	.copy )
}
\arguments{

\item{first_node}{first librdf_node node ("_p_librdf_node_s")}

\item{second_node}{second librdf_node node ("_p_librdf_node_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Compare two librdf_node objects for equality.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
