% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_node_get_blank_identifier}
\alias{librdf_node_get_blank_identifier}
\title{Get the blank node identifier as a UTF-8 encoded string.}
\usage{
librdf_node_get_blank_identifier ( node )
}
\arguments{

\item{node}{the node object ("_p_librdf_node_s")}

}
\value{
character
}
\description{
Get the blank node identifier as a UTF-8 encoded string.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
