% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_node_get_literal_value_as_latin1}
\alias{librdf_node_get_literal_value_as_latin1}
\title{Get the string literal value of the node as ISO Latin-1.}
\usage{
librdf_node_get_literal_value_as_latin1 ( node )
}
\arguments{

\item{node}{the node object ("_p_librdf_node_s")}

}
\value{
character
}
\description{
Get the string literal value of the node as ISO Latin-1.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
