% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_node_get_type}
\alias{librdf_node_get_type}
\title{Get the type of the node.}
\usage{
librdf_node_get_type ( node,
	.copy )
}
\arguments{

\item{node}{the node object ("_p_librdf_node_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Get the type of the node.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
