% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_node_get_uri}
\alias{librdf_node_get_uri}
\title{Get the URI for a node object.}
\usage{
librdf_node_get_uri ( node )
}
\arguments{

\item{node}{the node object ("_p_librdf_node_s")}

}
\value{
_p_librdf_uri_s
}
\description{
Get the URI for a node object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
