% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_node_is_literal}
\alias{librdf_node_is_literal}
\title{Check node is a literal.}
\usage{
librdf_node_is_literal ( node,
	.copy )
}
\arguments{

\item{node}{the node object ("_p_librdf_node_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Check node is a literal.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
