% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_node_to_string}
\alias{librdf_node_to_string}
\title{Format the node as a string in a debugging format.}
\usage{
librdf_node_to_string ( node )
}
\arguments{

\item{node}{the node object ("_p_librdf_node_s")}

}
\value{
character
}
\description{
Format the node as a string in a debugging format.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
