% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_parser_get_accept_header}
\alias{librdf_parser_get_accept_header}
\title{Get an HTTP Accept value for the parser.}
\usage{
librdf_parser_get_accept_header ( parser )
}
\arguments{

\item{parser}{parser ("_p_librdf_parser_s")}

}
\value{
character
}
\description{
Get an HTTP Accept value for the parser.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
