% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_parser_get_namespaces_seen_count}
\alias{librdf_parser_get_namespaces_seen_count}
\title{Get the number of namespaces seen during parsing}
\usage{
librdf_parser_get_namespaces_seen_count ( parser,
	.copy )
}
\arguments{

\item{parser}{librdf_parser object ("_p_librdf_parser_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Get the number of namespaces seen during parsing
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
