% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_parser_guess_name2}
\alias{librdf_parser_guess_name2}
\title{Get a parser name for content with type or identifier}
\usage{
librdf_parser_guess_name2 ( world,
	mime_type,
	buffer,
	identifier )
}
\arguments{

\item{world}{librdf_world object ("_p_librdf_world_s")}

\item{mime_type}{MIME type of syntax or NULL ("character")}

\item{buffer}{content buffer or NULL ("character")}

\item{identifier}{content identifier or NULL ("character")}

}
\value{
character
}
\description{
Get a parser name for content with type or identifier
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
