% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_parser_parse_as_stream}
\alias{librdf_parser_parse_as_stream}
\title{Parse a URI to a librdf_stream of statements.}
\usage{
librdf_parser_parse_as_stream ( parser,
	uri,
	inUriorNull )
}
\arguments{

\item{parser}{the parser ("_p_librdf_parser_s")}

\item{uri}{the URI to read ("_p_librdf_uri_s")}

\item{inUriorNull}{the base URI to use or NULL ("_p_librdf_uri_s")}

}
\value{
_p_librdf_stream_s
}
\description{
Parse a URI to a librdf_stream of statements.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
