% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_parser_parse_string_as_stream}
\alias{librdf_parser_parse_string_as_stream}
\title{Parse a string of content to a librdf_stream of statements.}
\usage{
librdf_parser_parse_string_as_stream ( parser,
	string,
	base_uri )
}
\arguments{

\item{parser}{the parser ("_p_librdf_parser_s")}

\item{string}{the string to parse ("character")}

\item{base_uri}{the base URI to use or NULL ("_p_librdf_uri_s")}

}
\value{
_p_librdf_stream_s
}
\description{
Parse a string of content to a librdf_stream of statements.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
