% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_parser_set_feature}
\alias{librdf_parser_set_feature}
\title{Set the value of a parser feature.}
\usage{
librdf_parser_set_feature ( parser,
	feature,
	value,
	.copy )
}
\arguments{

\item{parser}{librdf_parser object ("_p_librdf_parser_s")}

\item{feature}{librdf_uri feature property ("_p_librdf_uri_s")}

\item{value}{librdf_node feature property value ("_p_librdf_node_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Set the value of a parser feature.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
