% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_query_execute}
\alias{librdf_query_execute}
\title{Run the query on a model.}
\usage{
librdf_query_execute ( query,
	model )
}
\arguments{

\item{query}{librdf_query object ("_p_librdf_query")}

\item{model}{model to operate query on ("_p_librdf_model_s")}

}
\value{
_p_librdf_query_results
}
\description{
Run the query on a model.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
