% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_query_results_as_stream}
\alias{librdf_query_results_as_stream}
\title{Get a query result as an RDF graph in librdf_stream form}
\usage{
librdf_query_results_as_stream ( query_results )
}
\arguments{

\item{query_results}{librdf_query_results query_results ("_p_librdf_query_results")}

}
\value{
_p_librdf_stream_s
}
\description{
Get a query result as an RDF graph in librdf_stream form
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
