% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_query_results_finished}
\alias{librdf_query_results_finished}
\title{Find out if binding results are exhausted.}
\usage{
librdf_query_results_finished ( query_results,
	.copy )
}
\arguments{

\item{query_results}{librdf_query_results query results ("_p_librdf_query_results")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Find out if binding results are exhausted.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
