% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_query_results_get_boolean}
\alias{librdf_query_results_get_boolean}
\title{Get boolean query result.}
\usage{
librdf_query_results_get_boolean ( query_results,
	.copy )
}
\arguments{

\item{query_results}{librdf_query_results query_results ("_p_librdf_query_results")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Get boolean query result.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
