% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_query_results_is_graph}
\alias{librdf_query_results_is_graph}
\title{Test if librdf_query_results is RDF graph format.}
\usage{
librdf_query_results_is_graph ( query_results,
	.copy )
}
\arguments{

\item{query_results}{librdf_query_results object ("_p_librdf_query_results")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Test if librdf_query_results is RDF graph format.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
