% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_query_results_next}
\alias{librdf_query_results_next}
\title{Move to the next result.}
\usage{
librdf_query_results_next ( query_results,
	.copy )
}
\arguments{

\item{query_results}{librdf_query_results query results ("_p_librdf_query_results")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Move to the next result.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
