% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_query_set_limit}
\alias{librdf_query_set_limit}
\title{Set the query-specified limit on results.}
\usage{
librdf_query_set_limit ( query,
	limit,
	.copy )
}
\arguments{

\item{query}{librdf_query query object ("_p_librdf_query")}

\item{limit}{the limit on results, >=0 to set a limit, <0 to have no limit ("integer")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Set the query-specified limit on results.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
