% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_query_set_offset}
\alias{librdf_query_set_offset}
\title{Set the query-specified offset on results.}
\usage{
librdf_query_set_offset ( query,
	offset,
	.copy )
}
\arguments{

\item{query}{librdf_query query object ("_p_librdf_query")}

\item{offset}{offset for results, >=0 to set an offset, <0 to have no offset ("integer")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Set the query-specified offset on results.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
