% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_serializer_check_name}
\alias{librdf_serializer_check_name}
\title{Check if a serializer name is known}
\usage{
librdf_serializer_check_name ( world,
	name,
	.copy )
}
\arguments{

\item{world}{redland world object ("_p_librdf_world_s")}

\item{name}{name of serializer ("character")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Check if a serializer name is known
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
