% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_serializer_get_feature}
\alias{librdf_serializer_get_feature}
\title{Get the value of a serializer feature.}
\usage{
librdf_serializer_get_feature ( serializer,
	feature )
}
\arguments{

\item{serializer}{serializer object ("_p_librdf_serializer_s")}

\item{feature}{URI of feature ("_p_librdf_uri_s")}

}
\value{
_p_librdf_node_s
}
\description{
Get the value of a serializer feature.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
