% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_serializer_serialize_model_to_file}
\alias{librdf_serializer_serialize_model_to_file}
\title{Write a serialized librdf_model to a file.}
\usage{
librdf_serializer_serialize_model_to_file ( serializer,
	name,
	inUriOrNull,
	model,
	.copy )
}
\arguments{

\item{serializer}{the serializer ("_p_librdf_serializer_s")}

\item{name}{filename to serialize to ("character")}

\item{inUriOrNull}{the base URI to use (or NULL) ("_p_librdf_uri_s")}

\item{model}{the librdf_model model to use ("_p_librdf_model_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Write a serialized librdf_model to a file.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
