% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_serializer_serialize_model_to_string}
\alias{librdf_serializer_serialize_model_to_string}
\title{Write a serialized librdf_model to a string.
The returned string must be freed by the caller using librdf_free_memory().}
\usage{
librdf_serializer_serialize_model_to_string ( serializer,
	inUriOrNull,
	model )
}
\arguments{

\item{serializer}{the serializer ("_p_librdf_serializer_s")}

\item{inUriOrNull}{the base URI to use (or NULL) ("_p_librdf_uri_s")}

\item{model}{the librdf_model model to use ("_p_librdf_model_s")}

}
\value{
character
}
\description{
Write a serialized librdf_model to a string.
The returned string must be freed by the caller using librdf_free_memory().
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
