% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_serializer_serialize_stream_to_string}
\alias{librdf_serializer_serialize_stream_to_string}
\title{Write a librdf_stream to a string.}
\usage{
librdf_serializer_serialize_stream_to_string ( serializer,
	base_uri,
	stream )
}
\arguments{

\item{serializer}{the serializer ("_p_librdf_serializer_s")}

\item{base_uri}{the base URI to use (or NULL) ("_p_librdf_uri_s")}

\item{stream}{the librdf_stream stream to use ("_p_librdf_stream_s")}

}
\value{
character
}
\description{
Write a librdf_stream to a string.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
