% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_serializer_set_namespace}
\alias{librdf_serializer_set_namespace}
\title{Set a namespace URI/prefix mapping.}
\usage{
librdf_serializer_set_namespace ( serializer,
	nspace,
	prefix,
	.copy )
}
\arguments{

\item{serializer}{serializer object ("_p_librdf_serializer_s")}

\item{nspace}{URI of namespace or NULL ("_p_librdf_uri_s")}

\item{prefix}{prefix to use or NULL ("character")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Set a namespace URI/prefix mapping.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
