\name{librdf_short_copyright_string_get}
\alias{librdf_short_copyright_string_get}
\title{Return Redland librdf copyright string}
\usage{
librdf_short_copyright_string_get( .copy )
}
\arguments{

\item{.copy}{logical}

}
\value{
character
}
\description{
Return Redland librdf copyright string
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
