% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_statement_equals}
\alias{librdf_statement_equals}
\title{Check if two statements are equal.}
\usage{
librdf_statement_equals ( statement1,
	statement2,
	.copy )
}
\arguments{

\item{statement1}{first librdf_statement ("_p_librdf_statement_s")}

\item{statement2}{second librdf_statement ("_p_librdf_statement_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Check if two statements are equal.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
