% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_statement_get_predicate}
\alias{librdf_statement_get_predicate}
\title{Get the statement predicate.}
\usage{
librdf_statement_get_predicate ( statement )
}
\arguments{

\item{statement}{librdf_statement object ("_p_librdf_statement_s")}

}
\value{
_p_librdf_node_s
}
\description{
Get the statement predicate.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
