% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_statement_is_complete}
\alias{librdf_statement_is_complete}
\title{Check if statement is a complete and legal RDF triple.}
\usage{
librdf_statement_is_complete ( statement,
	.copy )
}
\arguments{

\item{statement}{librdf_statement object ("_p_librdf_statement_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Check if statement is a complete and legal RDF triple.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
