% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_statement_match}
\alias{librdf_statement_match}
\title{Match a statement against a 'partial' statement.}
\usage{
librdf_statement_match ( statement,
	partial_statement,
	.copy )
}
\arguments{

\item{statement}{statement ("_p_librdf_statement_s")}

\item{partial_statement}{statement with possible empty parts ("_p_librdf_statement_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Match a statement against a 'partial' statement.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
