% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_statement_set_object}
\alias{librdf_statement_set_object}
\title{Set the statement object.}
\usage{
librdf_statement_set_object ( statement,
	object )
}
\arguments{

\item{statement}{librdf_statement object ("_p_librdf_statement_s")}

\item{object}{librdf_node of object ("_p_librdf_node_s")}

}
\value{
void
}
\description{
Set the statement object.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
