% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_statement_set_predicate}
\alias{librdf_statement_set_predicate}
\title{Set the statement predicate.}
\usage{
librdf_statement_set_predicate ( statement,
	predicate )
}
\arguments{

\item{statement}{librdf_statement object ("_p_librdf_statement_s")}

\item{predicate}{librdf_node of predicate ("_p_librdf_node_s")}

}
\value{
void
}
\description{
Set the statement predicate.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
