% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_statement_to_string}
\alias{librdf_statement_to_string}
\title{Format the librdf_statement as a string.}
\usage{
librdf_statement_to_string ( statement )
}
\arguments{

\item{statement}{the statement ("_p_librdf_statement_s")}

}
\value{
character
}
\description{
Format the librdf_statement as a string.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
