% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_stream_end}
\alias{librdf_stream_end}
\title{Test if the stream has ended.}
\usage{
librdf_stream_end ( stream,
	.copy )
}
\arguments{

\item{stream}{librdf_stream object ("_p_librdf_stream_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Test if the stream has ended.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
