% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_stream_get_context}
\alias{librdf_stream_get_context}
\title{Get the context of the current object on the stream.}
\usage{
librdf_stream_get_context ( stream )
}
\arguments{

\item{stream}{the librdf_stream object ("_p_librdf_stream_s")}

}
\value{
_p_librdf_node_s
}
\description{
Get the context of the current object on the stream.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
