% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_stream_get_object}
\alias{librdf_stream_get_object}
\title{Get the current librdf_statement in the stream.}
\usage{
librdf_stream_get_object ( stream )
}
\arguments{

\item{stream}{librdf_stream object ("_p_librdf_stream_s")}

}
\value{
_p_librdf_statement_s
}
\description{
Get the current librdf_statement in the stream.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
