% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_uri_compare}
\alias{librdf_uri_compare}
\title{Compare two librdf_uri objects lexicographically.}
\usage{
librdf_uri_compare ( first_uri,
	second_uri,
	.copy )
}
\arguments{

\item{first_uri}{librdf_uri object 1 or NULL ("_p_librdf_uri_s")}

\item{second_uri}{librdf_uri object 2 or NULL ("_p_librdf_uri_s")}

\item{.copy}{NA}

}
\value{
integer
}
\description{
Compare two librdf_uri objects lexicographically.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
