% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_uri_to_string}
\alias{librdf_uri_to_string}
\title{Format the URI as a string.}
\usage{
librdf_uri_to_string ( uri )
}
\arguments{

\item{uri}{librdf_uri object ("_p_librdf_uri_s")}

}
\value{
character
}
\description{
Format the URI as a string.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
