\name{librdf_version_decimal_get}
\alias{librdf_version_decimal_get}
\title{Return Redland librdf copyright}
\usage{
librdf_version_decimal_get ( .copy )
}
\arguments{

\item{.copy}{logical}

}
\value{
integer
}
\description{
Return Redland librdf copyright
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
