% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_version_major}
\alias{librdf_version_major}
\title{Library major version number as a decimal integer.}
\usage{
librdf_version_major ( .copy )
}
\arguments{

\item{.copy}{NA}

}
\value{
integer
}
\description{
Library major version number as a decimal integer.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
