\name{librdf_version_release_get}
\alias{librdf_version_release_get}
\title{Return the Redland librdf release version number}
\usage{
librdf_version_release_get ( .copy )
}
\arguments{

\item{.copy}{logical}

}
\value{
integer
}
\description{
Return the Redland librdf release version number
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
