% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_version_string}
\alias{librdf_version_string}
\title{Library full version as a string.}
\usage{
librdf_version_string ( .copy )
}
\arguments{

\item{.copy}{NA}

}
\value{
character
}
\description{
Library full version as a string.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
