% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_world_get_feature}
\alias{librdf_world_get_feature}
\title{Get the value of a world feature.}
\usage{
librdf_world_get_feature ( world,
	feature )
}
\arguments{

\item{world}{librdf_world object ("_p_librdf_world_s")}

\item{feature}{librdf_uri feature property ("_p_librdf_uri_s")}

}
\value{
_p_librdf_node_s
}
\description{
Get the value of a world feature.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
