% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_world_open}
\alias{librdf_world_open}
\title{Open a created redland world environment.}
\usage{
librdf_world_open ( world )
}
\arguments{

\item{world}{redland world object ("_p_librdf_world_s")}

}
\value{
void
}
\description{
Open a created redland world environment.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
