% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{librdf_world_set_logger}
\alias{librdf_world_set_logger}
\title{Set the world log handling function.}
\usage{
librdf_world_set_logger ( world,
	user_data,
	log_handler )
}
\arguments{

\item{world}{redland world object ("_p_librdf_world_s")}

\item{user_data}{user data to pass to function ("_p_void")}

\item{log_handler}{pointer to the function ("_p_librdf_log_func")}

}
\value{
void
}
\description{
Set the world log handling function.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
