% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\docType{methods}
\name{initialize,Model-method}
\alias{initialize,Model-method}
\alias{Model-initialize}
\title{Constructor for a Model object.}
\usage{
\S4method{initialize}{Model}(.Object, world, storage, options)
}
\arguments{
\item{.Object}{a Node object}

\item{world}{a World object}

\item{storage}{a Storage object}

\item{options}{extra options for model initialization}
}
\value{
the World object
}
\description{
Constructor for a Model object.
}
