% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Node.R
\docType{class}
\name{Node-class}
\alias{Node-class}
\alias{getBlankNodeId,Node-method}
\alias{getNodeType,Node-method}
\title{A Redland Node, used to store one node in an RDF triple statement.}
\usage{
\S4method{getNodeType}{Node}(.Object)

\S4method{getBlankNodeId}{Node}(.Object)
}
\arguments{
\item{.Object}{a Node object}
}
\description{
A Node represents a RDF Resource, Property, Literal or an RDF blank Node.
}
\section{Methods (by generic)}{
\itemize{
\item \code{getNodeType}: 

\item \code{getBlankNodeId}: 
}}
\section{Slots}{

\describe{
\item{\code{librdf_node}}{A redland node object}
}}
\examples{
\dontrun{
# a blank node is created with a unique identifier generated by librdf
node <- new("Node", world)
# a blank node is created with a unique identifier generated by librdf
node <- new("Node", world, blank=NULL)
# a blank node is created with the user specified identifier, i.e. "_:id1"
node <- new("Node", world, blank="someid")
# a node type of 'literal' is created
node <- new("Node", world, literal="A Node Value")
# a Node type of 'resource' is created
node <- new("Node", world, uri="http://www.example.com")
}
}
\author{
Matthew Jones
}
\keyword{classes}

