% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser.R
\name{parseFileIntoModel}
\alias{parseFileIntoModel}
\title{Parse the contents of a file into a model}
\usage{
parseFileIntoModel(.Object, world, filePath, model, ...)
}
\arguments{
\item{.Object}{a Parser object}

\item{world}{a World object}

\item{filePath}{a file that contains the RDF content}

\item{model}{a Model object to parse the RDF content into}

\item{...}{baseUri a base URI (i.e. XML base) to apply to the model}
}
\description{
The contents of a the specified file are read and parsed into the initialized
Parser object
}
\details{
The parser factory name specified during initialization determines how the content is
parsed, for example, if 'rdfxml' was specified during parser initialization, then the parser
expects RDF/XML content as specified in the W3C recommendation (http://www.we3.org/TR/REC-rdf-syntax)
}

