% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parser.R
\docType{methods}
\name{initialize,Parser-method}
\alias{initialize,Parser-method}
\alias{Parser-initialize}
\title{Initialize a Parser object.}
\usage{
\S4method{initialize}{Parser}(.Object, world, name = "rdfxml",
  mimeType = "application/rdf+xml", typeUri = as.character(NA))
}
\arguments{
\item{.Object}{the Parser object}

\item{world}{a World object}

\item{name}{name of the parser factory to use}

\item{mimeType}{a mime type of the syntax of the model}

\item{typeUri}{a URI for the syntax of the model}
}
\value{
the Parser object
}
\description{
A Parser object is initialized for a specific RDF serialization.
}
\details{
The serialization format that are supported by
}
