% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\docType{class}
\name{Query-class}
\alias{Query}
\alias{Query-class}
\title{Query an RDF model}
\description{
The Query class is used to execuate a query on a Model object using the default query
language SPARQL. For more information, please refer to \url{http://librdf.org/rasqal/} for details on 
supported query languages.
}
\details{
A Query is executed using the executeQuery method, which returns a QueryResults object that
can be interated over the query solution sequence.
}
\section{Slots}{

\describe{
\item{\code{librdf_query}}{A redland query object}
}}
\section{Methods}{

\itemize{
  \item{\code{\link{Query-initialize}}}{: Initialize a Query object.}
  \item{\code{\link{executeQuery}}}{: Execute a query.}
  \item{\code{\link{setQueryResultLimit}}}{: Set limit on returned query results.}
  \item{\code{\link{getQueryResultLimit}}}{: Get the query result limit.}
  \item{\code{\link{freeParser}}}{: Free memory used by a librdf query.}
}
}
\examples{
\dontrun{
queryString <- 'PREFIX prov: <http://www.w3.org/ns/prov#> SELECT ?a ?c WHERE { ?a prov:Agent ?c . }'
query <- new("Query", world, queryString, base_uri=NULL, query_language="sparql", query_uri=NULL)
}
}
\references{
www.example.com
}
\seealso{
\code{\link{redland}}{: redland package}
}
\keyword{classes}

