% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\docType{methods}
\name{initialize,Query-method}
\alias{initialize,Query-method}
\alias{Query-initialize}
\title{Initialize the Query object.}
\usage{
\S4method{initialize}{Query}(.Object, world, querystring,
  base_uri = NULL, query_language = "sparql", query_uri = NULL)
}
\arguments{
\item{.Object}{the Query object}

\item{world}{a World object}

\item{querystring}{a query string for the language specified in 'query_language'}

\item{base_uri}{a URI to prepend to relative URI in the document}

\item{query_language}{the query language to execute the querystring with}

\item{query_uri}{a URI to prepend to terms in the query}
}
\value{
the Query object
}
\description{
Initialize the Query object.
}
