% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryResults.R
\docType{class}
\name{QueryResults-class}
\alias{QueryResults}
\alias{QueryResults-class}
\title{A Redland QueryResult object is used to inspect query results from a Query object.}
\description{
A Redland QueryResult object is used to inspect query results from a Query object.
}
\section{Slots}{

\describe{
\item{\code{librdf_query_results}}{A redland query object}
}}
\examples{
\dontrun{
query <- new("Query", world, queryString, base_uri=NULL, query_language="sparql", query_uri=NULL)
queryResult <- executeQuery(query, results)
}
}
\keyword{classes}

