% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryResults.R
\docType{methods}
\name{initialize,QueryResults-method}
\alias{initialize,QueryResults-method}
\alias{QueryResults-initialize}
\title{Initialize the QueryResults object.}
\usage{
\S4method{initialize}{QueryResults}(.Object, results)
}
\arguments{
\item{.Object}{the QueryResults object.}

\item{results}{a librdf query result}
}
\value{
the QueryResults object
}
\description{
The QueryResults object is initialized with the librdf query result from
return value of \code{'Query.execute()'}.
}
\details{
A QueryResults object is returned by the \code{Query.executeQuery()} method, so typically a user
does not initialize a QueryResult object by calling \code{new("QueryResult", ...)}
}
