\name{raptor_locator_byte}
\alias{raptor_locator_byte}
\title{Get the locator byte offset from locator.}
\usage{
raptor_locator_byte ( locator, .copy )
}
\arguments{

\item{locator}{raptor locator ("_p_raptor_locator")}

\item{.copy}{logical}

}
\value{
character
}
\description{
Get the locator byte offset from locator
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
