\name{raptor_locator_column}
\alias{raptor_locator_column}
\title{Get column number from locator}
\usage{
raptor_locator_column ( locator,
	.copy )
}
\arguments{

\item{locator}{raptor locator ("_p_raptor_locator")}

\item{.copy}{logical}

}
\value{
integer
}
\description{
Get column number from locator
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
