\name{raptor_locator_file}
\alias{raptor_locator_file}
\title{Get file name from locator.}
\usage{
raptor_locator_file ( locator )
}
\arguments{

\item{locator}{raptor locator ("_p_raptor_locator")}

}
\value{
character
}
\description{
Get file name from locator.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
