\name{raptor_locator_line}
\alias{raptor_locator_line}
\title{Get line number from locator.}
\usage{
raptor_locator_line ( locator, .copy )
}
\arguments{

\item{locator}{raptor locator ("_p_raptor_locator")}

\item{.copy}{logical}

}
\value{
integer
}
\description{
Get line number from locator.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
