\name{raptor_version_decimal}
\alias{raptor_version_decimal}
\title{Raptor version as a decimal number}
\usage{
raptor_version_decimal ( .copy )
}
\arguments{

\item{.copy}{logical}

}
\value{
integer
}
\description{
Raptor version as a decimal number
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
