\name{raptor_version_release_get}
\alias{raptor_version_release_get}
\title{Raptor library release.}
\usage{
raptor_version_release_get ( .copy )
}
\arguments{

\item{.copy}{logical}

}
\value{
integer
}
\description{
Get Raptor library release.
} 
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
