\name{raptor_version_string_get}
\alias{raptor_version_string_get}
\title{Get Raptor library version string.}
\usage{
raptor_version_string_get ( .copy )
}
\arguments{

\item{.copy}{logical}

}
\value{
character
}
\description{
Get Raptor library version string.
} 
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
