\name{rasqal_version_major_get}
\alias{rasqal_version_major_get}
\title{Get Rasqal major version number.}
\usage{
rasqal_version_major_get ( .copy ) 
}
\arguments{

\item{.copy}{logical}

}
\value{
integer
}
\description{
Get Rasqal major version number.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
