\name{rasqal_version_minor_get}
\alias{rasqal_version_minor_get}
\title{Get the Rasqal minor version number.}
\usage{
rasqal_version_minor_get ( .copy )
}
\arguments{

\item{.copy}{logical}

}
\value{
integer
}
\description{
Get the Rasqal minor version number.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries.
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
