\name{rasqal_version_string}
\alias{rasqal_version_string}
\title{Rasqal version as a string}
\usage{
rasqal_version_string ( .copy )
}
\arguments{

\item{.copy}{logical}

}
\value{
integer
}
\description{
Rasqal version as a string.
}
\references{
\url{http://librdf.org/docs}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
