% Generated by redland-bindings R program "createLibrdfDocs.R": do not edit by hand
\name{rasqal_world_set_warning_level}
\alias{rasqal_world_set_warning_level}
\title{Set the rasqal warning reporting level}
\usage{
rasqal_world_set_warning_level ( world,
	warning_level,
	.copy )
}

\arguments{


\item{world}{world ("_p_rasqal_world")}

\item{warning_level}{warning level 0..100 ("integer")}

\item{.copy}{NA}


}

\value{
integer
}
\description{
Set the rasqal warning reporting level
}
\references{
\url{https://librdf.org/docs/}
}
\seealso{
This R function is a wrapper function that directly calls the the Redland RDF C libraries. 
For more information about Redland RDF, view the online documentation indicated in the 'References' section.
}
