% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{redland-deprecated}
\alias{redland-deprecated}
\title{Deprecated Methods}
\description{
The following items are deprecated in this release of redland and will be
marked as Defunct and removed in a future version.
}
\section{These methods are deprecated}{

\itemize{
\item{\code{\link{getNextResult}}}{: Get the next result from a query.}
}
}

\keyword{internal}
