% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeNamespace_roclet.R
\name{roc_output.mergeNamespace}
\alias{roc_output.mergeNamespace}
\title{Roxygen output function that merges a base NAMESPACE file with the Roxygen dynamically created NAMSPACE file}
\usage{
roc_output.mergeNamespace(roclet, results, base_path, options = list(), check)
}
\arguments{
\item{roclet}{the currently running roclet}

\item{results}{the list of items to process that was generated by the roc_process.mergedNamespace function}

\item{base_path}{the base directory path of the package}

\item{options}{output options (not used)}

\item{check}{a check flag (not used)}
}
\description{
The 'roclet_output' function handles output of the results from the 'roc_process' function.
This function mergeds the NAMESPACE file created by the 'namespace' roclet with the list of Redland RDF
functions determined by the 'roc_process' function.
}

