% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeNamespace_roclet.R
\name{roclet_output.roclet_mergeNamespace}
\alias{roclet_output.roclet_mergeNamespace}
\title{Roxygen output function that merges a base NAMESPACE file with the Roxygen dynamically created NAMSPACE file}
\usage{
\method{roclet_output}{roclet_mergeNamespace}(x, results, base_path, ...)
}
\arguments{
\item{x}{the currently running roclet}

\item{results}{the list of items to process that was generated by the roc_process.mergedNamespace function}

\item{base_path}{the base directory path of the package}

\item{...}{additional parameters}
}
\description{
The 'roclet_output' function handles output of the results from the 'roc_process' function.
This function merges the NAMESPACE file created by the 'namespace' roclet with the list of Redland RDF
functions determined by the 'roc_process' function.
}
