% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeNamespace_roclet.R
\name{roclet_process.roclet_mergeNamespace}
\alias{roclet_process.roclet_mergeNamespace}
\title{Roxygen process function for the 'mergeNamespace' roclet}
\usage{
\method{roclet_process}{roclet_mergeNamespace}(roclet, partita, base_path,
  global_options = list())
}
\arguments{
\item{roclet}{the currently running roclet}

\item{partita}{a list of all .R files in the package being roxygenized (not used by this roclet)}

\item{base_path}{the top directory of the R package}

\item{global_options}{unused by this roclet}
}
\description{
This function is called by the Roxygen2 roxygenize function.
}
\details{
This function loads the Redland interface file and tests each loaded function to
see if it should be exported via the NAMESPACE file.
}
