% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Serializer.R
\docType{methods}
\name{initialize,Serializer-method}
\alias{initialize,Serializer-method}
\alias{Serializer-initialize}
\title{Construct a Serializer object.}
\usage{
\S4method{initialize}{Serializer}(.Object, world, name = "rdfxml",
  mimeType = "application/rdf+xml", typeUri = as.character(NA))
}
\arguments{
\item{.Object}{the Serializer object}

\item{world}{a World object}

\item{name}{name of a previously created serializer factory to use}

\item{mimeType}{a mime type of the syntax of the model}

\item{typeUri}{a URI for the syntax of the model}
}
\value{
the Serializer object
}
\description{
Construct a Serializer object.
}
