% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Serializer.R
\docType{methods}
\name{serializeToCharacter,Serializer,World,Model-method}
\alias{serializeToCharacter,Serializer,World,Model-method}
\title{Serialize a model to a character vector}
\usage{
\S4method{serializeToCharacter}{Serializer,World,Model}(.Object, world, model,
  baseUri = as.character(NA))
}
\arguments{
\item{.Object}{a Serializer object}

\item{world}{a World object}

\item{model}{a Model object}

\item{baseUri}{a URI to prepend to relative URIs in the document}
}
\description{
Serialize a model to a character vector
}

