% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Serializer.R
\name{serializeToCharacter}
\alias{serializeToCharacter}
\title{Serialize a model to a character vector}
\usage{
serializeToCharacter(.Object, world, model, ...)
}
\arguments{
\item{.Object}{a Serializer object}

\item{world}{a World object}

\item{model}{a Model object}

\item{...}{baseUri a URI to prepend to relative URIs in the document}
}
\value{
a character vector containing the serialized model
}
\description{
Serialize a model to a character vector
}

