% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Serializer.R
\name{serializeToFile}
\alias{serializeToFile}
\title{Serialize a model to a file}
\usage{
serializeToFile(.Object, world, model, filePath, ...)
}
\arguments{
\item{.Object}{a Serializer object}

\item{world}{a World object}

\item{model}{a Model object}

\item{filePath}{a file path that the serialized model will be written to}

\item{...}{baseUri a base URI to use for the serialization}
}
\value{
an integer containing the return status where non zero indicates an error occured during serialization
}
\description{
Serialize a model to a file
}

