% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Serializer.R
\docType{methods}
\name{setNameSpace,Serializer,World,character,character-method}
\alias{setNameSpace,Serializer,World,character,character-method}
\title{Set a namespace for the serializer}
\usage{
\S4method{setNameSpace}{Serializer,World,character,character}(.Object, world,
  namespace, prefix)
}
\arguments{
\item{.Object}{a Serializer object}

\item{world}{a World object}

\item{namespace}{the namespace to add to the serializer}

\item{prefix}{the namespace prefix to associate with the namespace}
}
\description{
Set a namespace for the serializer
}

