% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Query.R
\name{setQueryResultLimit}
\alias{setQueryResultLimit}
\title{Set limit on returned query results}
\usage{
setQueryResultLimit(.Object, limit)
}
\arguments{
\item{.Object}{a Query object}

\item{limit}{the result set limit. Specify a value >= to have a limit, or a value < 0 to have no limit.}
}
\description{
Set limit on returned query results
}

