% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Statement-class}
\alias{Statement-class}
\title{An RDF Statement object}
\description{
An RDF Statement object
}
\section{Slots}{

\describe{
\item{\code{librdf_statement}}{A redland storage object}
}}
\examples{
\dontrun{
stmt <- Statement(world, subject, predicate, object)
}
}
\author{
Peter Slaughter
}
\keyword{classes}

