% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Statement}
\alias{Statement}
\title{Construct a Statement object.}
\usage{
Statement(world, subject, predicate, object)
}
\arguments{
\item{world}{a World object}

\item{subject}{a Node object}

\item{predicate}{a Node object}

\item{object}{a Node object}
}
\value{
the Statement object
}
\description{
Construct a Statement object.
}

