% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storage.R
\docType{methods}
\name{initialize,Storage-method}
\alias{initialize,Storage-method}
\alias{Storage-initialize}
\title{Initialize a Storage object}
\usage{
\S4method{initialize}{Storage}(.Object, world, type = "hashes",
  name = "", options = "hash-type='memory'")
}
\arguments{
\item{.Object}{the Storage object}

\item{world}{the World object}

\item{type}{the Redland storage type}

\item{name}{storage instance name}

\item{options}{storage options}
}
\value{
the Storage object
}
\description{
Initialize a Storage object
}
\examples{
world <- new("World")
storage <- new("Storage", world, "hashes", name="", options="hash-type='memory'")
}
