% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{Storage,World,character,character,character-method}
\alias{Storage,World,character,character,character-method}
\title{Constructor for Storage object}
\usage{
\S4method{Storage}{World,character,character,character}(world, type, name,
  options)
}
\arguments{
\item{world}{the World object}

\item{type}{the Redland storage type}

\item{name}{storage instance name}

\item{options}{storage options}
}
\value{
the Storage object
}
\description{
Constructor for Storage object
}

