% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/redland.R
\docType{methods}
\name{[<-,ExternalReference-method}
\alias{[<-,ExternalReference-method}
\title{Do more stuff}
\usage{
\S4method{[}{ExternalReference}(x, i, j, ...) <- value
}
\arguments{
\item{x}{an ExternalReference}

\item{i}{a value}

\item{j}{a value}

\item{...}{some values}

\item{value}{a value}
}
\description{
We are doing lots of stuff now
}

