% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World.R
\docType{class}
\name{World-class}
\alias{World}
\alias{World-class}
\title{A Redland World object, used to initialize the Redland RDF library.}
\description{
A Redland World object, used to initialize the Redland RDF library.
}
\section{Slots}{

\describe{
\item{\code{librdf_world}}{A redland world object}
}}
\examples{
\dontrun{
world <- new("World")
}
}
\keyword{classes}

