% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World.R
\docType{methods}
\name{initialize,World-method}
\alias{initialize,World-method}
\alias{World-initialize}
\title{Initialize the World object.}
\usage{
\S4method{initialize}{World}(.Object)
}
\arguments{
\item{.Object}{the World object}
}
\value{
the World object
}
\description{
Initialize the World object.
}
