% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\docType{methods}
\name{addStatement}
\alias{addStatement}
\alias{addStatement,Model,Statement-method}
\title{Add a Statement object to the Model}
\usage{
addStatement(.Object, statement)

\S4method{addStatement}{Model,Statement}(.Object, statement)
}
\arguments{
\item{.Object}{a Model object}

\item{statement}{the Statement that will be added}
}
\description{
Add a Statement object to the Model
}
\examples{
world <- new("World")
storage <- new("Storage", world, "hashes", name="", options="hash-type='memory'")
model <- new("Model", world, storage, options="")
}
