% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addStatement}
\alias{addStatement}
\title{Add a Statement object to the Model}
\usage{
addStatement(.Object, statement)
}
\arguments{
\item{.Object}{a Model object}

\item{statement}{the Statement that will be added}
}
\description{
Add a Statement object to the Model
}

