% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\docType{methods}
\name{executeQuery}
\alias{executeQuery}
\alias{executeQuery,Query-method}
\title{Execute a query}
\usage{
executeQuery(.Object, model)

\S4method{executeQuery}{Query}(.Object, model)
}
\arguments{
\item{.Object}{a Query object}

\item{model}{a Model object containing the statements to query}
}
\value{
a QueryResults object
}
\description{
The initialize query is executed and the result is returned as a QueryResult object
}
