% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Parser.R
\name{freeParser}
\alias{freeParser}
\title{Free memory used by a librdf parser}
\usage{
freeParser(.Object)
}
\arguments{
\item{.Object}{a Node object}
}
\description{
Free memory used by a librdf parser
}
\details{
After freeNode is called, the Node object is no longer usable and should
be deleted  \code{"rm(nodeName)"} and a new object created.
}

