% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Query.R
\name{freeQuery}
\alias{freeQuery}
\title{Free memory used by a librdf query}
\usage{
freeQuery(.Object)
}
\arguments{
\item{.Object}{a Query object}
}
\description{
Free memory used by a librdf query
}
\details{
After this method is called, the Query object is no longer usable and should
be deleted \code{"rm(query)"} and a new object created.
}

