% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryResults.R
\docType{methods}
\name{freeQueryResults}
\alias{freeQueryResults}
\alias{freeQueryResults,QueryResults-method}
\title{Free memory used by a librdf query results}
\usage{
freeQueryResults(.Object)

\S4method{freeQueryResults}{QueryResults}(.Object)
}
\arguments{
\item{.Object}{a QueryResults object}
}
\description{
After this method is called, the QueryResults object is no longer usable and should
be deleted with \code{"rm(query)"}.
}
