% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/QueryResults.R
\name{freeQueryResults}
\alias{freeQueryResults}
\title{Free memory used by a librdf query results}
\usage{
freeQueryResults(.Object)
}
\arguments{
\item{.Object}{a QueryResults object}
}
\description{
Free memory used by a librdf query results
}
\details{
After this method is called, the QueryResults object is no longer usable and should
be deleted with \code{"rm(query)"}.
}

