% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Serializer.R
\docType{methods}
\name{freeSerializer}
\alias{freeSerializer}
\alias{freeSerializer,Serializer-method}
\title{Free memory used by a librdf serializer.}
\usage{
freeSerializer(.Object)

\S4method{freeSerializer}{Serializer}(.Object)
}
\arguments{
\item{.Object}{a Serializer object}
}
\description{
Free memory used by a librdf serializer.
}

