% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Statement.R
\name{freeStatement}
\alias{freeStatement}
\title{Free memory used by a librdf statement}
\usage{
freeStatement(.Object)
}
\arguments{
\item{.Object}{a Statement object}
}
\description{
Free memory used by a librdf statement
}
\details{
After this method is called, the Statement object is no longer usable and should
be deleted  \code{"rm(statement)"} and a new object created.
}

