% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Storage.R
\name{freeStorage}
\alias{freeStorage}
\title{Free memory used by a librdf storage object}
\usage{
freeStorage(.Object)
}
\arguments{
\item{.Object}{a Storage object to free memory for}
}
\description{
Free memory used by a librdf storage object
}
\details{
After this method is called, the Storage object is no longer usable and should
be deleted \code{"rm(storage)"} and a new object created.
}

