% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/World.R
\name{freeWorld}
\alias{freeWorld}
\title{Free memory used by a librdf world object}
\usage{
freeWorld(.Object)
}
\arguments{
\item{.Object}{a World object}
}
\description{
Free memory used by a librdf world object
}
\details{
After this method is called, the World object is no longer usable and should
be deleted \code{"rm(world)"} and a new object created.
}

