% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryResults.R
\docType{methods}
\name{getNextResult}
\alias{getNextResult}
\alias{getNextResult,QueryResults-method}
\title{Get the next query result.}
\usage{
getNextResult(.Object)

\S4method{getNextResult}{QueryResults}(.Object)
}
\arguments{
\item{.Object}{a QueryResults object}
}
\description{
The next query result is returned. .
}
