% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Node.R
\docType{methods}
\name{initialize,Node-method}
\alias{initialize,Node-method}
\title{Initialize a Node object.}
\usage{
\S4method{initialize}{Node}(.Object, world, literal, uri)
}
\arguments{
\item{.Object}{the Node object to be initialized}

\item{world}{a World object}

\item{literal}{a literal character value to be used as the vale of the node}

\item{uri}{a uri character value to be used as the value of the node}
}
\value{
the Node object
}
\description{
A Node has an associated type corresponding to the RDF component that it is representing. The list of possible
types is "resource", "literal" or "blank".
}
\details{
The url=' and 'literal=' arguments determine which type of Node is created. The Node type affects how the Node is processed
in serialization, for example a Node created with 'node1 <- new("Node", literal="http://www.example.com")' is processed
differently that a Node created with 'node1 <- new("Node", url="http://www.example.com")', with the former being processed
as an RDF literal and the latter processed as an RDF resource.
}

