% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Model.R
\docType{class}
\name{Model-class}
\alias{Model}
\alias{Model-class}
\alias{addStatement,Model,Statement-method}
\alias{freeModel,Model-method}
\title{A Redland Model object}
\usage{
\S4method{addStatement}{Model,Statement}(.Object, statement)

\S4method{freeModel}{Model}(.Object)
}
\arguments{
\item{.Object}{a Model object}

\item{statement}{the Statement that will be added}
}
\description{
A Model object is used to store the statements (triples) of an RDF model.
}
\details{
A Model may be created manually by creating \code{\link{Statement}} and adding
them to the Model using \code{\link{addStatement}}, or a Model may be read in from a
previously saved file using \code{\link{parseFileIntoModel}}. Once a Model is created,
it can be queried using \code{\link{Query}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{addStatement}: 

\item \code{freeModel}: 
}}
\section{Slots}{

\describe{
\item{\code{librdf_model}}{A redland model object}
}}
\examples{
\dontrun{
model <- new("Model", world, storage, options)
addStatement(model, statement)
}
}
\seealso{
View examples of creating models by viewing the \code{'redland_overview'} vignette: \code{'vignette("redland_overview")'}
}
\keyword{classes}

