% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Model-class}
\alias{Model-class}
\title{An RDF Model object.}
\description{
An RDF Model object.
}
\section{Slots}{

\describe{
\item{\code{librdf_model}}{A redland model object}
}}
\examples{
\dontrun{
model <- new("Model", world, storage, options)
addStatement(model, statement)
}
}
\author{
Peter Slaughter
}
\keyword{classes}

