% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\docType{class}
\name{Model-class}
\alias{Model}
\alias{Model-class}
\title{A Redland Model object}
\description{
A Model object is used to store the statements (triples) of an RDF model.
}
\details{
A Model may be created manually by creating \code{\link{Statement}} and adding
them to the Model using \code{\link{addStatement}}, or a Model may be read in from a
previously saved file using \code{\link{parseFileIntoModel}}. Once a Model is created,
it can be queried using \code{\link{Query}}.
}
\section{Slots}{

\describe{
\item{\code{librdf_model}}{A redland model object}
}}
\examples{
\dontrun{
model <- new("Model", world, storage, options)
addStatement(model, statement)
}
}
\seealso{
View examples of creating models by viewing the \code{'redland_overview'} vignette: \code{'vignette("redland_overview")'}
}
\keyword{classes}

