% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Node-class}
\alias{Node-class}
\title{A Redland Node, used to store one node in an RDF triple statement.}
\description{
A Node represents a RDF Resource, Property, or Literal.
}
\section{Slots}{

\describe{
\item{\code{librdf_node}}{A redland node object}
}}
\examples{
\dontrun{
node <- new("Node", world, "A Node Value")
}
}
\author{
Matthew Jones
}
\keyword{classes}

