% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Node.R
\docType{class}
\name{Node-class}
\alias{Node-class}
\title{A Redland Node, used to store one node in an RDF triple statement.}
\description{
A Node represents a RDF Resource, Property, Literal or an RDF blank Node.
}
\section{Slots}{

\describe{
\item{\code{librdf_node}}{A redland node object}
}}
\examples{
\dontrun{
node <- new("Node", world)     # a blank node is created
node <- new("Node", world, "A Node Value")     # a Node type of 'literal' is created
node <- new("Node", world, literal="A Node Value")     # a node type of 'literal' is created
node <- new("Node", world, uri="http://www.example.com")   # a Node type of 'resource' is created
}
}
\author{
Matthew Jones
}
\keyword{classes}

