% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Parser.R
\docType{class}
\name{Parser-class}
\alias{Parser-class}
\alias{freeParser,Parser-method}
\alias{parseFileIntoModel,Parser,World,character,Model-method}
\title{An RDF Parser object}
\usage{
\S4method{parseFileIntoModel}{Parser,World,character,Model}(.Object, world,
  filePath, model, baseUri = as.character(NA))

\S4method{freeParser}{Parser}(.Object)
}
\arguments{
\item{.Object}{a Parser object}

\item{world}{a World object}

\item{filePath}{a file that contains the RDF content}

\item{model}{a Model object to parse the RDF content into}

\item{baseUri}{a base URI (i.e. XML base) to apply to the model}
}
\description{
The Parser class provides methods to parse RDF content into a Redland
RDF model.
}
\section{Methods (by generic)}{
\itemize{
\item \code{parseFileIntoModel}: #' 

\item \code{freeParser}: 
}}
\section{Slots}{

\describe{
\item{\code{librdf_parser}}{A redland parser object}
}}
\examples{
\dontrun{
parser <- new("Parser", world, name, mimeType, typeUri)
parserIntoModel(parser, uri, baseUri, model)
}
}
\author{
Peter Slaughter
}
\keyword{classes}

