% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Query.R
\docType{class}
\name{Query-class}
\alias{Query}
\alias{Query-class}
\alias{executeQuery,Query-method}
\alias{freeQuery,Query-method}
\alias{getQueryResultLimit,Query-method}
\alias{setQueryResultLimit,Query-method}
\title{A Redland Query object, used to query an RDF model}
\usage{
\S4method{executeQuery}{Query}(.Object, model)

\S4method{setQueryResultLimit}{Query}(.Object, limit)

\S4method{getQueryResultLimit}{Query}(.Object)

\S4method{freeQuery}{Query}(.Object)
}
\arguments{
\item{.Object}{a Query object}

\item{model}{a Model object}

\item{limit}{max number of results to return}
}
\description{
The Query class is used to execuate a query on a Model object using the default query
language SPARQL. For more information, please refer to \url{http://librdf.org/rasqal/} for details on
supported query languages.
}
\details{
A Query is executed using the executeQuery method, which returns a QueryResults object that
can be interated over the query solution sequence.
}
\section{Methods (by generic)}{
\itemize{
\item \code{executeQuery}: 

\item \code{setQueryResultLimit}: 

\item \code{getQueryResultLimit}: 

\item \code{freeQuery}: 
}}
\section{Slots}{

\describe{
\item{\code{librdf_query}}{A redland query object}
}}
\examples{
\dontrun{
queryString <- 'PREFIX prov: <http://www.w3.org/ns/prov#> SELECT ?a ?c WHERE { ?a prov:Agent ?c . }'
query <- new("Query", world, queryString, base_uri=NULL, query_language="sparql", query_uri=NULL)
}
}
\references{
www.example.com
}
\keyword{classes}

