% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Query.R
\docType{class}
\name{Query-class}
\alias{Query-class}
\alias{executeQuery,Query-method}
\alias{freeQuery,Query-method}
\alias{getQueryResultLimit,Query-method}
\alias{setQueryResultLimit,Query-method}
\title{A Query object is used to query an RDF model}
\usage{
\S4method{executeQuery}{Query}(.Object, model)

\S4method{setQueryResultLimit}{Query}(.Object, limit)

\S4method{getQueryResultLimit}{Query}(.Object)

\S4method{freeQuery}{Query}(.Object)
}
\description{
A Query object is used to query an RDF model
}
\section{Methods (by generic)}{
\itemize{
\item \code{executeQuery}: 

\item \code{setQueryResultLimit}: 

\item \code{getQueryResultLimit}: 

\item \code{freeQuery}: 
}}
\section{Slots}{

\describe{
\item{\code{librdf_query}}{A redland query object}
}}
\examples{
\dontrun{
queryString <- 'PREFIX prov: <http://www.w3.org/ns/prov#> SELECT ?a ?c WHERE { ?a prov:Agent ?c . }'
query <- new("Query", world, queryString, base_uri=NULL, query_language="sparql", query_uri=NULL)
}
}
\author{
Scott Chamberlain, Peter Slaughter
}
\keyword{classes}

