% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryResults.R
\docType{class}
\name{QueryResults-class}
\alias{QueryResults-class}
\alias{QueryResults}
\title{A Redland QueryResults object is used to inspect query results from a Query object.}
\description{
The QueryResults object contains the RDF statements that were returned from
a query on an RDF model.
}
\section{Slots}{

\describe{
\item{\code{librdf_query_results}}{A redland query object}
}}

\section{Methods}{

\itemize{
\item{\code{\link{QueryResults-initialize}}}{: Initialize a QueryResults object.}
\item{\code{\link{freeQueryResults}}}{: Free memory used by a librdf query result.}
}
}

\seealso{
\code{\link{redland}}{: redland package}
}
\keyword{classes}
