% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/QueryResults.R
\docType{class}
\name{QueryResults-class}
\alias{QueryResults-class}
\alias{freeQueryResults,QueryResults-method}
\alias{getNextResult,QueryResults-method}
\title{QueryResult is used to retrieve query results from a Query object.}
\usage{
\S4method{getNextResult}{QueryResults}(.Object)

\S4method{freeQueryResults}{QueryResults}(.Object)
}
\description{
QueryResult is used to retrieve query results from a Query object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{getNextResult}: 

\item \code{freeQueryResults}: 
}}
\section{Slots}{

\describe{
\item{\code{librdf_query}}{A redland query object}
}}
\examples{
\dontrun{
query <- new("Query", world, queryString, base_uri=NULL, query_language="sparql", query_uri=NULL)
queryResult <- executeQuery(query, results)
}
}
\author{
Peter Slaughter
}
\keyword{classes}

