% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/redland-package.R
\docType{package}
\name{redland}
\alias{redland}
\alias{redland-package}
\title{Create, query and write RDF graphs.}
\description{
The redland R package provides methods to create, query and write information stored in the
Resource Description Framework (RDF).
}
\author{
Matthew B. Jones (NCEAS) and Peter Slaughter (NCEAS)
}

