% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeNamespace_roclet.R
\name{roclet_process.roclet_mergeNamespace}
\alias{roclet_process.roclet_mergeNamespace}
\title{Roxygen process function for the 'mergeNamespace' roclet}
\usage{
\method{roclet_process}{roclet_mergeNamespace}(x, blocks, env, base_path,
  global_options = list())
}
\arguments{
\item{x}{the currently running roclet}

\item{blocks}{the documentation blocks}

\item{env}{the current env}

\item{base_path}{the top directory of the R package}

\item{global_options}{unused by this roclet}
}
\description{
This function is called by the Roxygen2 roxygenize function.
}
\details{
This function loads the Redland interface file and tests each loaded function to
see if it should be exported via the NAMESPACE file.
}
