% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Serializer.R
\docType{class}
\name{Serializer-class}
\alias{Serializer-class}
\title{An RDF Serializer object}
\description{
An RDF Serializer object
}
\section{Slots}{

\describe{
\item{\code{librdf_serializer}}{A redland statement object}
}}
\examples{
\dontrun{
serializer <- new("Serializer", world, factoryName, mimeType, typeUri)
setNameSpace(serializer, world, namespace, prefix)
RDF <- serializeToCharacter(serializer, world, model, baseUri)
}
}
\author{
Peter Slaughter
}
\keyword{classes}

