% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Serializer.R
\docType{methods}
\name{serializeToCharacter}
\alias{serializeToCharacter}
\alias{serializeToCharacter,Serializer,World,Model-method}
\title{Serialize a model to a character vector.}
\usage{
serializeToCharacter(.Object, world, model, ...)

\S4method{serializeToCharacter}{Serializer,World,Model}(.Object, world,
  model, baseUri = as.character(NA))
}
\arguments{
\item{.Object}{a Serializer object}

\item{world}{a World object}

\item{model}{a Model object}

\item{...}{Additional parameters}

\item{baseUri}{a URI to prepend to relative URIs in the document}
}
\value{
a character vector containing the serialized model
}
\description{
Serialize a model to a character vector.
}
