% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storage.R
\docType{class}
\name{Storage-class}
\alias{Storage-class}
\alias{Storage}
\title{A Redland Storage object}
\description{
A Redland Storage object
}
\section{Slots}{

\describe{
\item{\code{librdf_storage}}{A redland storage object}

\item{\code{type}}{the storage type to create, i.e. "hashes", "mysql", "postgresql", ...}
}}

\section{Methods}{

\itemize{
\item{\code{\link{Storage-initialize}}}{: Initialize a Storage object}
\item{\code{\link{freeStorage}}}{: Free memory used by a librdf storage object}
}
}

\examples{
world <- new("World")
storage <- new("Storage", world, "hashes", name="", options="hash-type='memory'")
}
\seealso{
\code{\link{redland}}{: redland package}
}
\keyword{classes}
