% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Storage.R
\docType{class}
\name{Storage-class}
\alias{Storage}
\alias{Storage-class}
\alias{freeStorage,Storage-method}
\title{A Redland Storage object}
\usage{
\S4method{freeStorage}{Storage}(.Object)
}
\arguments{
\item{.Object}{a Storage object}
}
\description{
A Redland Storage object
}
\section{Methods (by generic)}{
\itemize{
\item \code{freeStorage}: 
}}
\section{Slots}{

\describe{
\item{\code{librdf_storage}}{A redland storage object}

\item{\code{type}}{the storage type to create, i.e. "hashes", "mysql", "postgresql", ...}
}}
\examples{
\dontrun{
storage <- new("Storage", world, "hashes", name="", options="hash-type='memory'")
}
}
\keyword{classes}

