% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Storage-class}
\alias{Storage-class}
\title{A Redland Storage object}
\description{
A Redland Storage object
}
\section{Slots}{

\describe{
\item{\code{librdf_storage}}{A redland storage object}
}}
\examples{
\dontrun{
storage <- new("Storage", world, "hashes", name="", options="hash-type='memory'")
}
}
\author{
Peter Slaughter
}
\keyword{classes}

