\docType{methods}
\name{[,ExternalReference-method}
\alias{[,ExternalReference-method}
\title{Subset a list of ExternalReferences}
\usage{
\S4method{[}{ExternalReference}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a list of ExternalReferences}

\item{i}{row subscript}

\item{j}{column subscript}

\item{...}{additional arguments}

\item{drop}{a logical}
}
\description{
Subset a list of ExternalReferences
}

