% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/redland.R
\docType{methods}
\name{[,ExternalReference-method}
\alias{[,ExternalReference-method}
\title{Do some stuff}
\usage{
\S4method{[}{ExternalReference}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{an ExternalReference}

\item{i}{a value}

\item{j}{a value}

\item{...}{some values}

\item{drop}{a logical}
}
\description{
this method does some stuff
}

