\docType{methods}
\name{[<-,ExternalReference-method}
\alias{[<-,ExternalReference-method}
\title{Assign values in a list of ExternalReferences}
\usage{
\S4method{[}{ExternalReference}(x, i, j, ...) <- value
}
\arguments{
\item{x}{a list of ExternalReferences}

\item{i}{row subscript}

\item{j}{column subscript}

\item{...}{additional arguments}

\item{value}{a value to assign}
}
\description{
Assign values in a list of ExternalReferences
}

