% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/redland.R
\docType{methods}
\name{[<-,ExternalReference-method}
\alias{[<-,ExternalReference-method}
\title{Assign values in a list of ExternalReferences}
\usage{
\S4method{[}{ExternalReference}(x, i, j, ...) <- value
}
\arguments{
\item{x}{a list of ExternalReferences}

\item{i}{row subscript}

\item{j}{column subscript}

\item{...}{additional arguments}

\item{value}{a value to assign}
}
\description{
Assign values in a list of ExternalReferences
}

