% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/World.R
\docType{class}
\name{World-class}
\alias{World}
\alias{World-class}
\alias{freeWorld,World-method}
\title{A Redland World object, used to initialize the Redland RDF library.}
\usage{
\S4method{freeWorld}{World}(.Object)
}
\arguments{
\item{.Object}{a World object}
}
\description{
A Redland World object, used to initialize the Redland RDF library.
}
\section{Methods (by generic)}{
\itemize{
\item \code{freeWorld}: 
}}
\section{Slots}{

\describe{
\item{\code{librdf_world}}{A redland world object}
}}
\examples{
\dontrun{
world <- new("World")
}
}
\keyword{classes}

