% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{World-class}
\alias{World-class}
\title{A Redland World, used to initialize the Redland RDF library.}
\description{
A Redland World, used to initialize the Redland RDF library.
}
\section{Slots}{

\describe{
\item{\code{world}}{A redland world object}
}}
\examples{
\dontrun{
world <- new("World")
}
}
\author{
Matthew Jones
}
\keyword{classes}

